+++ src/components/main/Notice.tsx
@@ -1,7 +1,6 @@
 import { color } from "@/styles/color";
 import { font } from "@/styles/font";
 import RightArrowIcon from "@/components/common/Icon/RightArrow";
-import NoticeItem from "./NoticeItem";
 import styled from "styled-components";
 import { useRouter } from "next/navigation";
 import Link from "@/components/common/Link";
@@ -19,7 +18,11 @@ const Notice = () => {
       </Link>
       <NoticeList>
         {data.map((item) => (
-          <NoticeItem key={item.id} id={item.id} title={item.title} />
+          <NoticeItem>
+            <NoticeItemTitle onClick={() => router.push(`/notice/${item.id}`)}>
+              {item.title}
+            </NoticeItemTitle>
+          </NoticeItem>
         ))}
       </NoticeList>
     </StyledNotice>
@@ -45,4 +48,23 @@ const NoticeList = styled.div`
   width: 100%;
   height: 100%;
   overflow-y: auto;
+`;
+
+const NoticeItem = styled.div`
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 33.3%;
+  padding: 0px 16px;
+  border-bottom: 1px solid ${color.gray300};
+`;
+
+const NoticeItemTitle = styled.a`
+  ${font.p1}
+  color: ${color.gray750};
+  cursor: pointer;
+  // 일정 길이 넘어가면 ... 처리
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 `;
